Created and Configurated
This initial commit establishes the basic structure of the SoftClever.com.br project and includes the following changes:

Development Environment Setup:

Utilized Create React App to create the React development environment.
Installed and configured Tailwind CSS to facilitate quick and responsive styling of the project.
Integration of Linters:

Added ESLint to ensure code consistency and quality.
Configured Prettier to maintain a uniform and readable code style.
These modifications establish an initial foundation for the project, providing the necessary tools for efficient development and maintenance of the code.