ADD: Page ASoftClever
Added content and styling for the Home page

Updated the Home component (Home.tsx) to include content about "A Soft Clever" and its values.
Utilized the Header, Menu, Footer, and Whatsapp components for consistent layout and navigation.
Structured the content to showcase the company's values and services.
Included an image to complement the content.
Ensured responsiveness for different screen sizes.