ADD: Contato Page and Fixes
Implemented a dedicated contact page to facilitate communication with users. This page includes input fields for users to submit their name, email, company name, phone number, and a message. It serves as a platform for users to reach out and initiate communication with the company.

Contact Page Creation: Developed a new page titled "Contato" with input fields for name, email, company name, phone number, and message. Users can utilize these fields to send inquiries or messages to the company.
Route Update: Added a route to the contact page in the main.tsx file, enabling seamless navigation to the newly created contact page.